+++ crates/swc_typescript/tests/fast_dts_deno.rs
